requirements.txt
flask==3.0.0
flask-mysqldb==1.0.1
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-limiter==3.5.0
scikit-learn==1.3.2
nltk==3.8.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
pandas==2.1.4
numpy==1.26.2 
